+++ website/app.py
@@ -1,34 +1,32 @@
 # -*- coding: utf-8 -*-
 
-import os
 import importlib
-from collections import OrderedDict
 import json
+import os
+from collections import OrderedDict
 
 import django
-
-from modularodm import storage
 from werkzeug.contrib.fixers import ProxyFix
+
 import framework
+import website.models
+from framework.addons.utils import render_addon_capabilities
 from framework.flask import app, add_handlers
 from framework.logging import logger
-from framework.mongo import set_up_storage
-from framework.addons.utils import render_addon_capabilities
-from framework.sentry import sentry
 from framework.mongo import handlers as mongo_handlers
+from framework.mongo import set_up_storage
+from framework.postcommit_tasks import postcommit_handlers as postcommit_handlers
+from framework.sentry import sentry
 from framework.tasks import handlers as task_handlers
-from framework.tasks import postcommit_handlers as postcommit_handlers
 from framework.transactions import handlers as transaction_handlers
-
-import website.models
-from website.routes import make_url_map
+from modularodm import storage
 from website.addons.base import init_addon
-from website.project.model import ensure_schemas, Node
 from website.project.licenses import ensure_licenses
+from website.project.model import ensure_schemas, Node
+from website.routes import make_url_map
+
+
 # This import is necessary to set up the archiver signal listeners
-from website.archiver import listeners  # noqa
-from website.mails import listeners  # noqa
-from api.caching import listeners  # noqa
 
 def build_js_config_files(settings):
     with open(os.path.join(settings.STATIC_FOLDER, 'built', 'nodeCategories.json'), 'wb') as fp:
